added socket in like/dislike

+++ src/app/modules/user-profile/blog-management/blog-management.component.ts
@@ -4,6 +4,7 @@ import { UserProfileService } from 'src/app/core/services/user-profile.service';
 import { WriteBlogService } from 'src/app/core/services/write-blog.service';
 import { PATHS } from 'src/app/common/constants';
 import { Router } from '@angular/router';
+import swal from 'sweetalert2'
 
 
 @Component({
@@ -40,6 +41,11 @@ constructor(private userService:UserProfileService,private blogService:WriteBlog
   }
 
   blockUser(){
+    swal.fire(
+      'Good job!',
+      'User Blocked',
+      'warning'
+    )
     console.log("Blocked User")
   }
   deleteUser(){
@@ -59,11 +65,32 @@ constructor(private userService:UserProfileService,private blogService:WriteBlog
     
   }
   deleteBlog(id:string){
-    this.blogService.deleteBlog(id).subscribe();
+    swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result?.isConfirmed) {
+        this.blogService.deleteBlog(id).subscribe();
+        this.getAllBlogs();
+        swal.fire(
+          'Deleted!',
+          'Your Blog has been deleted.',
+          'success'
+        )
+      }
+    })
+     
+  }
+  getAllBlogs(){
     this.blogService.getUserBlogs().subscribe((res:any)=>{
       this.blogData=res.data;
-      console.log(res.data)
-    })
+      console.log(res.data, "gg")
+    }) 
   }
   
 }
