+++ src/app/modules/user-profile/user-blogs/user-blogs.component.ts
@@ -2,6 +2,7 @@ import { Component} from '@angular/core';
 import { Router } from '@angular/router';
 import { PATHS } from 'src/app/common/constants';
 import { WriteBlogService } from 'src/app/core/services/write-blog.service';
+import swal from 'sweetalert2'
 
 
 @Component({
@@ -17,26 +18,44 @@ export class UserBlogsComponent {
   ngOnInit(){
     this.blogService.getUserBlogs().subscribe((res:any)=>{
       this.allBlogs=res.data;
-      console.log(res.data)
+     
     })
   }
   edit(id:string){
     let blogData;
      this.allBlogs.find((res:any)=>{
-      console.log(res.blog.blogId)
+     
       if(res.blog.blogId===id){
         blogData= res;
-    this.router.navigateByUrl(PATHS.MAIN.BLOG_WRITE,{state:{data:blogData}})
+    this.router.navigateByUrl(PATHS.MAIN?.BLOG_WRITE,{state:{data:blogData}})
 
       }
      })
     
   }
   delete(id:string){
-    this.blogService.deleteBlog(id).subscribe();
-    this.blogService.getUserBlogs().subscribe((res:any)=>{
-      this.allBlogs=res.data;
-      console.log(res.data)
+    swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.blogService.deleteBlog(id).subscribe();
+        this.blogService.getUserBlogs().subscribe((res:any)=>{
+          this.allBlogs=res.data;
+        })
+        swal.fire(
+          'Deleted!',
+          'Your Blog has been deleted.',
+          'success'
+        )
+      }
     })
+
+   
   }
 }
